Added file input, moved more functions to common
