+++ src/common.c
@@ -226,6 +226,81 @@ int check_file(char *file){
   }
 }
 
+int check_exec(const char *object)
+{
+  if (access(object, X_OK) == 0) {
+    return 1;
+  } else {
+    return 0;
+  }
+}
+
+int check_last_char(const char *str, const char *chk)
+{
+  if (!strcmp(&str[strlen(str) - 1], chk)){
+    return 1;
+  } else {
+    return 0;
+  }
+}
+
+int check_first_char(const char *str, const char *chk)
+{
+  if (str[0] == chk[0]){
+    return 1;
+  } else {
+    return 0;
+  }
+}
+
+// Credit for the following function must go to this guy:
+// https://stackoverflow.com/a/779960
+char *str_replace(char *orig, char *rep, char *with) {
+    char *result; // the return string
+    char *ins;    // the next insert point
+    char *tmp;    // varies
+    int len_rep;  // length of rep (the string to remove)
+    int len_with; // length of with (the string to replace rep with)
+    int len_front; // distance between rep and end of last rep
+    int count;    // number of replacements
+
+    // sanity checks and initialization
+    if (!orig || !rep)
+        return NULL;
+    len_rep = strlen(rep);
+    if (len_rep == 0)
+        return NULL; // empty rep causes infinite loop during count
+    if (!with)
+        with = "";
+    len_with = strlen(with);
+
+    // count the number of replacements needed
+    ins = orig;
+    for (count = 0; tmp = strstr(ins, rep); ++count) {
+        ins = tmp + len_rep;
+    }
+
+    tmp = result = malloc(strlen(orig) + (len_with - len_rep) * count + 1);
+
+    if (!result)
+        return NULL;
+
+    // first time through the loop, all the variable are set correctly
+    // from here on,
+    //    tmp points to the end of the result string
+    //    ins points to the next occurrence of rep in orig
+    //    orig points to the remainder of orig after "end of rep"
+    while (count--) {
+        ins = strstr(orig, rep);
+        len_front = ins - orig;
+        tmp = strncpy(tmp, orig, len_front) + len_front;
+        tmp = strcpy(tmp, with) + len_with;
+        orig += len_front + len_rep; // move to next "end of rep"
+    }
+    strcpy(tmp, orig);
+    return result;
+}
+
 void clear_workspace()
 {
   size_t line_count = 1;
