+++ src/sfmenus.h
@@ -1,1 +1,2 @@
 void show_file_inputs();
+void show_file_file_input();
